restructuring error html python file to more closely resemble the html structure
