+++ csss-site/src/elections/views/create_context/nominee_links/create_or_update_nominee/create_context_for_create_or_update_nominee__nominee_links_html.py
@@ -1,15 +1,18 @@
 import json
 import logging
 
-from csss.views.context_creation.create_context_for_html_snippet_for_general_error_validations import \
+from csss.views.context_creation.error_htmls.create_context_for_html_snippet_for_general_error_validations import \
     create_context_for_html_snippet_for_general_error_validations_html
 from elections.models import NomineeLink
 from elections.views.Constants import NOMINEE__NAME__HTML_NAME, ELECTION_HUMAN_FRIENDLY_NAME__HTML_NAME
-from elections.views.create_context.nominee_links.create_or_update_nominee.create_context_for_form__nominee_links_html import \
+from elections.views.create_context.nominee_links.\
+    create_or_update_nominee.create_context_for_form__nominee_links_html import \
     create_context_for_form__nominee_links_html
-from elections.views.create_context.nominee_links.create_or_update_nominee.create_context_for_view_saved_nominee_info_html import \
+from elections.views.create_context.nominee_links.create_or_update_nominee.\
+    create_context_for_view_saved_nominee_info_html import \
     create_context_for_view_saved_nominee_info_html
-from elections.views.create_context.nominee_links.create_or_update_nominee.js_functions.on_load_js_functions.create_context_for_main_function__nominee_links_html import \
+from elections.views.create_context.nominee_links.create_or_update_nominee.js_functions.on_load_js_functions.\
+    create_context_for_main_function__nominee_links_html import \
     create_context_for_main_function__nominee_links_html
 from elections.views.create_context.nominee_links.utils.make_context_value_serializable_to_json import \
     make_json_serializable_context_dictionary
