+++ csss-site/src/elections/views/create_context/webform/create_context_for_create_election__webform_html.py
@@ -1,4 +1,4 @@
-from csss.views.context_creation.create_context_for_html_snippet_for_general_error_validations import \
+from csss.views.context_creation.error_htmls.create_context_for_html_snippet_for_general_error_validations import \
     create_context_for_html_snippet_for_general_error_validations_html
 from elections.views.ElectionModelConstants import ELECTION_JSON_KEY__DATE, ELECTION_JSON_WEBFORM_KEY__TIME, \
     ELECTION_JSON_KEY__ELECTION_TYPE, ELECTION_JSON_KEY__WEBSURVEY, ELECTION_JSON_KEY__NOMINEES
